backend/requirements.txt
asgiref==3.6.0
astroid==2.15.3
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.1
defusedxml==0.7.1
dill==0.3.6
Django==2.2.16
django-filter==2.4.0
django-stubs-ext==0.8.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-extra-fields==3.4.1
gunicorn==20.1.0
idna==3.4
importlib-metadata==1.7.0
itypes==1.2.0
Jinja2==3.1.2
lazy-object-proxy==1.9.0
MarkupSafe==2.1.2
mccabe==0.7.0
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.1
pathspec==0.11.1
Pillow==9.5.0
platformdirs==3.2.0
psycopg2-binary==2.8.6
pycparser==2.21
PyJWT==2.6.0
pylint-plugin-utils==0.7
python-dotenv==0.21.1
python3-openid==3.2.0
pytz==2022.7.1
reportlab==3.6.12
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.4.0
sqlparse==0.4.3
tomli==2.0.1
tomlkit==0.11.7
typed-ast==1.5.4
types-pytz==2023.3.0.0
types-PyYAML==6.0.12.9
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.15
wrapt==1.15.0
zipp==3.15.0
